% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-methods.R
\name{get_sampleIds}
\alias{get_sampleIds}
\title{Extract Sample IDs from list of files}
\usage{
get_sampleIds(directory = ".", pattern_def = ".tsv", split_def = "[.]")
}
\arguments{
\item{directory}{the path to directory that contains the sample files}

\item{pattern_def}{the pattern string, the part after the end of the sample IDs part}

\item{split_def}{the character that is used to split words in file names}
}
\value{
A list of 2 elements, (1)the sample list & (2)column position with sample IDs in names
}
\description{
Extract Sample IDs from list of files
}
