% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO-methods.R
\name{read_pathoscope}
\alias{read_pathoscope}
\title{Read Pathoscope output}
\usage{
read_pathoscope(dir_file = ".", s_list = NULL,
  smeta_file = "sample_metadata.csv", is.dir = 1, pattern0 = "",
  pattern1 = ".tsv", splitPoint = "[.]")
}
\arguments{
\item{dir_file}{path for the directory with all sample files}

\item{s_list}{a list of all the sample IDs involved in the experiment}

\item{smeta_file}{a csv/tsv file containing the sample meta data}

\item{is.dir}{a flag parameter, if the "dir_file" parameter is a directory(1) or a file (0)}

\item{pattern0}{& pattern1 these are the pattern of strings before and after the sample IDs}

\item{splitPoint}{the character that is used to split words in file names}
}
\value{
A list of "Count data matrix" and "Sample metadata data frame" and the "TaxIDs for the OTUs"
}
\description{
Read Pathoscope output
}
