% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid2lineage.R
\name{uid2lineage}
\alias{uid2lineage}
\title{Download lineage for NCBI taxonomy ID}
\usage{
uid2lineage(x, ranks = c("superkingdom", "phylum", "class", "order", "family",
  "genus", "species"))
}
\arguments{
\item{x}{Vector of taxa names (character) or IDs (character or numeric) to 
query. Passed to \code{\link[taxize]{classification}}.}

\item{ranks}{Vector of rank labels for lineage.}
}
\value{
A data.frame with the lineage for each supplied taxonomy ID.
}
\description{
Uses taxize to download the lineage for given taxon IDs
}
\examples{
taxids <-c(39313, 2, 48381, 95953, 4923, 11111, 111111111)
lineage.df <- uid2lineage(taxids)

}
